requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio

dvc==3.24.0
dvc-gdrive==2.20.0
langchain==0.0.316
dill==0.3.7
transformers==4.34.0
sentence-transformers==2.2.2
chromadb==0.4.13
datasets==2.14.5
pandas==2.1.1
matplotlib==3.8.0
trl==0.7.4
peft==0.5.0
accelerate==0.23.0
python-dotenv==1.0.0
wandb==0.15.12
scikit-learn==1.3.2
openai==0.28.1
lark==1.1.8
pypdf==3.17.1
lxml==4.9.3
#bitsandbytes==0.41.1 # only works for linux, macos ....why....(sad emoji here)

#windows workaround
--extra-index-url https://jllllll.github.io/bitsandbytes-windows-webui
--prefer-binary
bitsandbytes